requirements.txt
fastapi
uvicorn
pydantic
email-validator
bcrypt
python-multipart
httpx
PyJWT
python-dotenv

